+++ docs/changelog.md
@@ -18,6 +18,7 @@
 * Changed `otter generate autograder` to `otter generate`
 * Removed `otter generate token` as all interaction with `otter.generate.token.APIClient` can be handled elsewhere
 * Added intercell seeding for R Jupyter Notebooks
+* Added `ValueError` on unexpected config in `otter.assign.assignment.Assignment`
 
 **v1.1.6:**